requirements.txt
streamlit
transformers
torch
sentencepiece
pdfplumber
python-docx
pytesseract
pillow
sqlalchemy
alembic
uvicorn
python-multipart
